requirements.txt
argcomplete==1.9.3
atomicwrites==1.2.1
attrs==18.2.0
boto3==1.9.51
botocore==1.12.51
Cerberus==1.2
certifi==2018.10.15
cfn-flip==1.1.0.post1
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
django-extensions==2.1.4
django-heroku==0.3.1
djangorestframework==3.9.0
docutils==0.14
durationpy==0.5
future==0.16.0
gunicorn==19.9.0
hjson==3.0.1
idna==2.7
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
more-itertools==4.3.0
placebo==0.8.2
pluggy==0.8.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py==1.7.0
pytest==4.0.1
pytest-django==3.4.4
python-dateutil==2.6.1
python-decouple==3.1
python-slugify==1.2.4
pytz==2018.7
PyYAML==3.13
requests==2.20.1
s3transfer==0.1.13
six==1.11.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.3.4
Unidecode==1.0.23
urllib3==1.24.1
Werkzeug==0.14.1
whitenoise==4.1.2
wsgi-request-logger==0.4.6
zappa==0.47.1
zappa-django-utils==0.4.0
